+++ CHANGELOG.md
@@ -6,9 +6,11 @@ All notable changes to this project will be documented in this file based on the
 * Removed support for Elastica as Elasticsearch client
 * Removed adapters completely (Repository is now directly using `\Elasticsearch\Client`)
 * Removed `ElasticaQuery`
+* Removed method `ElasticsearchRepository::deleteIndex()` in favor of `IndexManager::deleteIndex()`
 ### Bugfixes
 ### Added
 * `postSave` and `postDelete()` hooks for repositories
+* Index management features via `IndexManager`
 ### Improvements
 ### Deprecated